requirements.txt
carpi-commons==0.1.1
carpi-daemoncommons==0.2.0
carpi-gpsdaemon==0.2.3
carpi-obddaemon==0.2.1
carpi-redisdatabus==0.2.0
carpi-settings==0.1.2
obd==0.7.0
Pint==0.7.2
pyserial==3.4
pytz==2018.7
redis==3.0.1
tzlocal==1.5.1
